+++ NEWS.md
@@ -1,3 +1,5 @@
 ## 1.0.0 2019-10-21
 * Initial module setup
+* Added event manager
 * Added transport layer to publish event to consumer services
+* Added mapping manager